requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
bootstrap4==0.1.0
Django==4.0.3
django-bootstrap-v5==1.0.11
django-bootstrap5==21.3
django-ckeditor==6.3.0
django-js-asset==2.0.0
soupsieve==2.3.1
sqlparse==0.4.2
style==1.1.0
tzdata==2022.1
update==0.0.1
